requirements_python3.6.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.4.1
requests==2.21.0
SQLAlchemy==1.2.16
urllib3==1.24.1
vk==2.0.2
Werkzeug==0.14.1
